Added FERA eligibility
